requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
db-dtypes==1.0.2
fastapi==0.63.0
fastapi-utils==0.2.1
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-bigquery==3.3.0
google-cloud-bigquery-storage==2.14.1
google-cloud-core==2.3.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
idna==3.3
mangum==0.11.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
proto-plus==1.20.6
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.39
starlette==0.13.6
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.13.4
